Refactor how the solver models operations; support NewEntry operation (#66)
The motivating factor here is the desire to support NewEntry operations;
however, considering how to do this prompted a larger refactor, which
aligns the terminology in the solver closer to that in the core POD2
library, and replaces the previous trait-based system with a simple set
of functions, one for each possible operation.

These now also have tests, which is a big improvement over the lack of
tests for `PredicateHandler`s.