Refactor app navigation and data stores (#77)
This PR does three things:

- Divides the app state store into "mini-apps"
- Separates out navigation between apps and within apps
- Adds the `immer` library for making app-state updates

### Mini-apps

These are:
- POD Collection explorer
- Documents
- POD editing
- FrogCrypto

This should result in minimal changes, e.g. instead of doing things
like:
```
const { frogTimeout } = useAppStore();
```
We can instead do:
```
const { frogTimeout } = useFrogCrypto();
```

It's not a huge change but it gives each app its own storage area.

### Navigation

Previously the app had a global navigation state (currentView and
previousView), which allowed for limited "back" navigation in certain
contexts. This was pretty hacky and ad-hoc.

Now the app has a single global state variable which determines which
mini-app is visible. Internally, mini-apps can have their own more
detailed history if necessary. Users can switch between mini-apps using
the sidebar.

The Documents app now has a browser-style back/forward navigation, which
allows the user to go back/forward between the Documents list,
individual viewed documents, and the drafts/publish screens. The other
apps don't have enough internal complexity to need this kind of
navigation, but they are now free to have their own internal navigation
systems as appropriate.

### Immer

In short, the `immer` package makes updates to nested JavaScript objects
easier.

The longer version is that React re-renders components whenever their
inputs have changed. This includes "props", the direct inputs to a
component, but also things like the data stores that a component is
subscribed to.

For each component subscribed to a data store, React will check if the
subset of the data store that the component has subscribed to has
changed. However, if you have a JavaScript object graph, there's no
built-in way to know if some nested property of an object has changed,
since the object's identity is unchanged.

If we have a reference to `store.object.child`, and then the `prop`
field on that object changes to some other value, the identity of
`store.object.child` is unaffected - it's still the same object, but
with a new value for one of its fields.

React state management libraries solve this by treating state objects as
effectively immutable: you can't just change the value on a property,
you have to replace the object with a new (shallow) copy with one or
more updated properties. React will then notice that the new object has
a different identity to the old one, and will update any subscribers as
appropriate.

However, this kind of "rewrite the object graph while shallow-copying
the things that haven't changed" code is tedious to write. Immer solves
this by providing a proxy object which wraps the state, and which lets
you write natural "mutable-style" object updates, e.g.:

```typescript
navigateToDocument: (id: number) => {
  set((state) => {
    const newRoute: DocumentRoute = { type: "document-detail", id };
    const history = state.documents.browsingHistory;
    // Trim forward history
    history.stack.splice(history.currentIndex + 1);
    // Add new route
    history.stack.push(newRoute);
    history.currentIndex = history.stack.length - 1;
  });
},
```

Here we call the `set` function, passing in a callback which will
receive the `state` proxy object. We then update the properties
directly, and Immer handles the rewriting of the object graph, so that
React can see which objects have had their properties changed.