Real-time markdown preview and synced scrolling (#89)
How this works:

On each keypress, we send a notification to a worker thread, which
parses the new state of the Markdown document and works out what
changed. It then sends back a message to the main thread, which can
surgically update the parts of the preview that have changed.

We use a shared array buffer to keep a sequence ID for edits, so that
the worker thread can know if it has processed all of the pending
events; it will only run the parser if it has received all of the
pending events.