Initial branch-and-bound strategy for picking optimal proof
We maintain a lower bound based on the smallest number of PODs required for a
solution seen so far, allowing us to prune branches which will inevitably
require more PODs. There's probably more optimisation needed here, and
definitely more tests. We can also adopt a strategy of returning immediately
if we find a solution that can be completed in one POD, since we will never
find a better one.